Revert "Remove A/B test from shady pie"
